# patient-service/src/main/java/com/sm/patientservice/model/Insurance.java
package com.sm.patientservice.model;

import jakarta.persistence.Embeddable;
import lombok.Data;

@Embeddable
@Data
public class Insurance {
    
    private String provider;
    private String policyNumber;
   

}